Joined Students/Rooms w Student_Rooms, added Users
In the JavaScript for the Students table, I also inserted a studentId and a roomId into the Student_Rooms table so that the two are now connected. I additionally am pushing the Users table which contains information on users. As such, all of our SQL tables have now been generated with several rows of data in each!