Comments file incorporating djm feedback
Same as last pull request with updated 500 comments and filled database but this one has the suggestions from Professor Malan incorporated. I would have committed the changes to the previous pull request but I've been having lots of trouble in my terminal pushing changes to pull requests because git continues to tell me that there are merge conflicts but I'm unable to get the file in which I can resolve them in terminal.