Added Username/Password to Users
I added 200 sample Users to the Users table, which now contains userName and password fields with a maximum of 20 characters. This allows us to work with the login page by adding the proposed username and password to the Users table if one registers and also running through the Users table to see if a given user is registered.